+++ Marlin/component.mk
@@ -3,4 +3,4 @@
 #
 # (Uses default behaviour of compiling all source files in directory, adding 'include' to include path.)
 
-COMPONENT_SRCDIRS = . src src/config src/config/default  src/core src/feature src/feature/bedlevel src/feature/bedlevel/abl src/feature/bedlevel/mbl src/feature/bedlevel/ubl src/feature/dac src/feature/digipot src/feature/leds src/gcode src/gcode/bedlevel src/gcode/bedlevel/abl src/gcode/bedlevel/mbl src/gcode/bedlevel/ubl src/gcode/calibrate src/gcode/config src/gcode/control src/gcode/eeprom src/gcode/feature src/gcode/feature/advance src/gcode/feature/baricuda src/gcode/feature/camera src/gcode/feature/caselight src/gcode/feature/clean src/gcode/feature/digipot src/gcode/feature/filwidth src/gcode/feature/fwretract src/gcode/feature/i2c src/gcode/feature/leds src/gcode/feature/mixing src/gcode/feature/pause src/gcode/feature/snmm src/gcode/feature/trinamic src/gcode/geometry src/gcode/host src/gcode/lcd src/gcode/motion src/gcode/probe src/gcode/scara src/gcode/sdcard src/gcode/stats src/gcode/temperature src/gcode/units src/HAL src/HAL/HAL_ESP src/inc src/lcd src/lcd/dogm src/lcd/language src/libs src/module src/module/thermistor src/pins src/sd
+COMPONENT_SRCDIRS = . src src/config src/config/default  src/core src/feature src/feature/bedlevel src/feature/bedlevel/abl src/feature/bedlevel/mbl src/feature/bedlevel/ubl src/feature/dac src/feature/digipot src/feature/leds src/gcode src/gcode/bedlevel src/gcode/bedlevel/abl src/gcode/bedlevel/mbl src/gcode/bedlevel/ubl src/gcode/calibrate src/gcode/config src/gcode/control src/gcode/eeprom src/gcode/feature src/gcode/feature/advance src/gcode/feature/baricuda src/gcode/feature/camera src/gcode/feature/caselight src/gcode/feature/clean src/gcode/feature/digipot src/gcode/feature/filwidth src/gcode/feature/fwretract src/gcode/feature/i2c src/gcode/feature/leds src/gcode/feature/mixing src/gcode/feature/pause src/gcode/feature/snmm src/gcode/feature/trinamic src/gcode/geometry src/gcode/host src/gcode/lcd src/gcode/motion src/gcode/probe src/gcode/scara src/gcode/sdcard src/gcode/stats src/gcode/temperature src/gcode/units src/HAL src/HAL/HAL_ESP32 src/inc src/lcd src/lcd/dogm src/lcd/language src/libs src/module src/module/thermistor src/pins src/sd
