+++ Marlin/src/HAL/HAL_ESP32/SanityCheck_ESP32.h
@@ -19,33 +19,3 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-
-#ifndef HAL_SPI_PINS_H_
-#define HAL_SPI_PINS_H_
-
-#ifdef ARDUINO_ARCH_SAM
-  #include "HAL_DUE/spi_pins.h"
-
-#elif defined(IS_32BIT_TEENSY)
-  #include "HAL_TEENSY35_36/spi_pins.h"
-
-#elif defined(__AVR__)
-  #include "HAL_AVR/spi_pins.h"
-
-#elif defined(TARGET_LPC1768)
-  #include "HAL_LPC1768/spi_pins.h"
-
-#elif defined(__STM32F1__)
-  #include "HAL_STM32F1/spi_pins.h"
-
-#elif defined(STM32F7)
-  #include "HAL_STM32F7/spi_pins.h"
-
-#elif defined(ARDUINO_ARCH_ESP32)
-  #include "HAL_ESP32/spi_pins.h"
-
-#else
-  #error "Unsupported Platform!"
-#endif
-
-#endif // HAL_SPI_PINS_H_
