+++ Marlin/src/pins/pins_ESP32.h
@@ -54,11 +54,6 @@
 #define E0_ENABLE_PIN      131
 #define E0_CS_PIN          -1
 
-// #define E1_STEP_PIN        36
-// #define E1_DIR_PIN         34
-// #define E1_ENABLE_PIN      30
-// #define E1_CS_PIN          44
-
 //
 // Temperature Sensors
 //
@@ -110,30 +105,6 @@
 
 #define HEATER_BED_PIN   34
 
-// #if ENABLED(IS_RAMPS_EFB)                      // Hotend, Fan, Bed
-//   #define FAN_PIN        RAMPS_D9_PIN
-//   #define HEATER_BED_PIN RAMPS_D8_PIN
-// #elif ENABLED(IS_RAMPS_EEF)                    // Hotend, Hotend, Fan
-//   #define HEATER_1_PIN   RAMPS_D9_PIN
-//   #define FAN_PIN        RAMPS_D8_PIN
-// #elif ENABLED(IS_RAMPS_EEB)                    // Hotend, Hotend, Bed
-//   #define HEATER_1_PIN   RAMPS_D9_PIN
-//   #define HEATER_BED_PIN RAMPS_D8_PIN
-// #elif ENABLED(IS_RAMPS_EFF)                    // Hotend, Fan, Fan
-//   #define FAN_PIN        RAMPS_D9_PIN
-//   #define FAN1_PIN       RAMPS_D8_PIN
-// #elif ENABLED(IS_RAMPS_SF)                     // Spindle, Fan
-//   #define FAN_PIN        RAMPS_D8_PIN
-// #else                                          // Non-specific are "EFB" (i.e., "EFBF" or "EFBE")
-//   #define FAN_PIN        RAMPS_D9_PIN
-//   #define HEATER_BED_PIN RAMPS_D8_PIN
-//   #if HOTENDS == 1
-//     #define FAN1_PIN     MOSFET_D_PIN
-//   #else
-//     #define HEATER_1_PIN MOSFET_D_PIN
-//   #endif
-// #endif
-
 //
 // Misc. Functions
 //
@@ -147,161 +118,3 @@
 #define FIL_RUNOUT_PIN      4
 
 #define PS_ON_PIN          12
-
-//
-// LCD / Controller
-//
-#if ENABLED(ULTRA_LCD)
-
-  #if ENABLED(REPRAPWORLD_GRAPHICAL_LCD)
-    #define LCD_PINS_RS     49 // CS chip select /SS chip slave select
-    #define LCD_PINS_ENABLE 51 // SID (MOSI)
-    #define LCD_PINS_D4     52 // SCK (CLK) clock
-  #elif ENABLED(NEWPANEL) && ENABLED(PANEL_ONE)
-    #define LCD_PINS_RS 40
-    #define LCD_PINS_ENABLE 42
-    #define LCD_PINS_D4 65
-    #define LCD_PINS_D5 66
-    #define LCD_PINS_D6 44
-    #define LCD_PINS_D7 64
-  #else
-    #define LCD_PINS_RS 16
-    #define LCD_PINS_ENABLE 17
-    #define LCD_PINS_D4 23
-    #define LCD_PINS_D5 -1
-    #define LCD_PINS_D6 27
-    #define LCD_PINS_D7 29
-    #if DISABLED(NEWPANEL)
-      #define BEEPER_PIN -1
-      // Buttons are attached to a shift register
-      // Not wired yet
-      //#define SHIFT_CLK 38
-      //#define SHIFT_LD 42
-      //#define SHIFT_OUT 40
-      //#define SHIFT_EN 17
-    #endif
-  #endif
-
-  #if ENABLED(NEWPANEL)
-
-    #if ENABLED(REPRAP_DISCOUNT_SMART_CONTROLLER)
-      #define BEEPER_PIN 37
-
-      #define BTN_EN1 31
-      #define BTN_EN2 -1
-      #define BTN_ENC 35
-
-      #define SD_DETECT_PIN 49
-      #define KILL_PIN 41
-
-      #if ENABLED(BQ_LCD_SMART_CONTROLLER)
-        #define LCD_BACKLIGHT_PIN 39
-      #endif
-
-    #elif ENABLED(REPRAPWORLD_GRAPHICAL_LCD)
-      #define BTN_EN1 64
-      #define BTN_EN2 59
-      #define BTN_ENC 63
-      #define SD_DETECT_PIN 42
-    #elif ENABLED(LCD_I2C_PANELOLU2)
-      #define BTN_EN1 47  // reverse if the encoder turns the wrong way.
-      #define BTN_EN2 43
-      #define BTN_ENC 32
-      #define LCD_SDSS 53
-      #define SD_DETECT_PIN -1
-      #define KILL_PIN 41
-    #elif ENABLED(LCD_I2C_VIKI)
-      #define BTN_EN1 22  // reverse if the encoder turns the wrong way.
-      #define BTN_EN2 7   // http://files.panucatt.com/datasheets/viki_wiring_diagram.pdf
-                          // tells about 40/42.
-                          // 22/7 are unused on RAMPS_14. 22 is unused and 7 the SERVO0_PIN on RAMPS_13.
-      #define BTN_ENC -1
-      #define LCD_SDSS 53
-      #define SD_DETECT_PIN 49
-    #elif ENABLED(VIKI2) || ENABLED(miniVIKI)
-      #define BEEPER_PIN       -1
-
-      // Pins for DOGM SPI LCD Support
-      #define DOGLCD_A0        44
-      #define DOGLCD_CS        45
-      #define LCD_SCREEN_ROT_180
-
-      #define BTN_EN1          22
-      #define BTN_EN2           7
-      #define BTN_ENC          39
-
-      #define SDSS             53
-      #define SD_DETECT_PIN    -1  // Pin 49 for display sd interface, 72 for easy adapter board
-
-      #define KILL_PIN         31
-
-      #define STAT_LED_RED_PIN 32
-      #define STAT_LED_BLUE_PIN 35
-
-    #elif ENABLED(ELB_FULL_GRAPHIC_CONTROLLER)
-      #define BTN_EN1 35  // reverse if the encoder turns the wrong way.
-      #define BTN_EN2 37
-      #define BTN_ENC 31
-      #define SD_DETECT_PIN 49
-      #define LCD_SDSS 53
-      #define KILL_PIN 41
-      #define BEEPER_PIN 23
-      #define DOGLCD_CS 29
-      #define DOGLCD_A0 27
-      #define LCD_BACKLIGHT_PIN -1
-    #elif ENABLED(MINIPANEL)
-      #define BEEPER_PIN 42
-      // Pins for DOGM SPI LCD Support
-      #define DOGLCD_A0  44
-      #define DOGLCD_CS  66
-      #define LCD_BACKLIGHT_PIN 65 // backlight LED on A11/D65
-      #define SDSS   53
-
-      #define KILL_PIN 64
-      // GLCD features
-      //#define LCD_CONTRAST 190
-      // Uncomment screen orientation
-      //#define LCD_SCREEN_ROT_90
-      //#define LCD_SCREEN_ROT_180
-      //#define LCD_SCREEN_ROT_270
-      // The encoder and click button
-      #define BTN_EN1 40
-      #define BTN_EN2 63
-      #define BTN_ENC 59
-      // not connected to a pin
-      #define SD_DETECT_PIN 49
-
-    #else
-
-      // Beeper on AUX-4
-      #define BEEPER_PIN -1
-
-      // buttons are directly attached using AUX-2
-      #if ENABLED(REPRAPWORLD_KEYPAD)
-        #define BTN_EN1 64 // encoder
-        #define BTN_EN2 59 // encoder
-        #define BTN_ENC 63 // enter button
-        #define SHIFT_OUT 40 // shift register
-        #define SHIFT_CLK 44 // shift register
-        #define SHIFT_LD 42 // shift register
-      #elif ENABLED(PANEL_ONE)
-        #define BTN_EN1 59 // AUX2 PIN 3
-        #define BTN_EN2 63 // AUX2 PIN 4
-        #define BTN_ENC 49 // AUX3 PIN 7
-      #else
-        #define BTN_EN1 37
-        #define BTN_EN2 35
-        #define BTN_ENC 31 // the click
-      #endif
-
-      #if ENABLED(G3D_PANEL)
-        #define SD_DETECT_PIN 49
-        #define KILL_PIN 41
-      #else
-        //#define SD_DETECT_PIN -1 // Ramps doesn't use this
-      #endif
-
-    #endif
-  #endif // NEWPANEL
-
-#endif // ULTRA_LCD
